made the pastRunSection its own component
